celery-potato/requirements-experiment.txt
celery[gevent]==5.1.2
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-celery==0.26b1
opentelemetry-semantic-conventions==0.26b1
git+https://github.com/lonewolf3739/opentelemetry-python.git@bugfix-issue-2266#egg=opentelemetry-sdk&subdirectory=opentelemetry-sdk
